feat(version): support --version flag
This adds support for the `--version` CLI flag. This outputs the version
number for both shx and the bundled version of ShellJS. This adds a test
to strictly assert the output format, so that this can be reliably
parsed if necessary.

This also reorganizes the structure of some test cases (without changing
their logic).

Fixes #111
Test: case: 'supports --version'