feat(version): support --version flag (#138)
This adds support for the `--version` CLI flag. This outputs the version
number for both shx and the bundled version of ShellJS. This adds a test
to strictly assert the output format, so that this can be reliably
parsed if necessary.

This also reorganizes the structure of some test cases (without changing
their logic).

Lastly, this changes the file modes for 2 files. These were previously
unnecessarily executable.

Fixes #111
Test: case: 'supports --version'